# Refactored vertical_slash_index.cu for performance improvement