@Starmys 2024-10-09
Thank you for your contribution! Could you provide your CUDA version so we can check if this is related to compiler optimization ?